refactor: extract toText helper
