+++ src/index.ts
@@ -1,92 +1,6 @@
 // rrule-temporal.ts
 import { Temporal } from "@js-temporal/polyfill";
 
-// ---------------------------------------------------------------------------
-// Helper utilities for the toText() feature
-// ---------------------------------------------------------------------------
-const WEEKDAY_NAMES = [
-  "Monday",
-  "Tuesday",
-  "Wednesday",
-  "Thursday",
-  "Friday",
-  "Saturday",
-  "Sunday",
-];
-
-const MONTH_NAMES = [
-  "January",
-  "February",
-  "March",
-  "April",
-  "May",
-  "June",
-  "July",
-  "August",
-  "September",
-  "October",
-  "November",
-  "December",
-];
-
-export type DateFormatter = (
-  year: number,
-  month: string,
-  day: number
-) => string;
-
-const defaultDateFormatter: DateFormatter = (
-  year: number,
-  month: string,
-  day: number
-) => `${month} ${day}, ${year}`;
-
-function ordinal(n: number): string {
-  const abs = Math.abs(n);
-  const suffix =
-    abs % 10 === 1 && abs % 100 !== 11
-      ? "st"
-      : abs % 10 === 2 && abs % 100 !== 12
-      ? "nd"
-      : abs % 10 === 3 && abs % 100 !== 13
-      ? "rd"
-      : "th";
-  return n < 0 ? `last` : `${abs}${suffix}`;
-}
-
-function list(
-  arr: (string | number)[],
-  mapFn: (x: string | number) => string = (x) => `${x}`,
-  final = "and"
-): string {
-  const mapped = arr.map(mapFn);
-  if (mapped.length === 1) return mapped[0]!;
-  return (
-    mapped.slice(0, -1).join(", ") + ` ${final} ` + mapped[mapped.length - 1]
-  );
-}
-
-function formatByDayToken(tok: string | number): string {
-  if (typeof tok === "number") return tok.toString();
-  const m = tok.match(/^([+-]?\d+)?(MO|TU|WE|TH|FR|SA|SU)$/);
-  if (!m) return tok;
-  const ord = m[1] ? parseInt(m[1], 10) : 0;
-  const weekdayMap: { [key: string]: number } = {
-    MO: 0,
-    TU: 1,
-    WE: 2,
-    TH: 3,
-    FR: 4,
-    SA: 5,
-    SU: 6,
-  };
-  const idx = weekdayMap[m[2] as keyof typeof weekdayMap];
-  const name = WEEKDAY_NAMES[idx!];
-  if (ord === 0) return name!;
-  if (ord === -1) return `last ${name}`;
-  return `${ordinal(ord)} ${name}`;
-}
-
 // Allowed frequency values
 type Freq =
   | "YEARLY"
@@ -361,24 +275,6 @@ function parseRRuleString(
 //   return ord ? `${nth(ord)} ${name}` : name;
 // }
 
-function formatTime(hour: number, minute = 0): string {
-  const hr12 = ((hour + 11) % 12) + 1;
-  const ampm = hour < 12 ? "AM" : "PM";
-  if (minute) {
-    return `${hr12}:${String(minute).padStart(2, "0")} ${ampm}`;
-  }
-  return `${hr12} ${ampm}`;
-}
-
-function tzAbbreviation(zdt: Temporal.ZonedDateTime): string {
-  const parts = new Intl.DateTimeFormat("en-US", {
-    timeZone: zdt.timeZoneId,
-    timeZoneName: "short",
-    hour: "numeric",
-  }).formatToParts(new Date(zdt.epochMilliseconds));
-  const tzPart = parts.find((p) => p.type === "timeZoneName");
-  return tzPart?.value || zdt.timeZoneId;
-}
 
 export class RRuleTemporal {
   private tzid: string;
@@ -1642,98 +1538,6 @@ export class RRuleTemporal {
     return [dtLine, `RRULE:${parts.join(";")}`].join("\n");
   }
 
-  /**
-   * Convert this rule to a human readable English description.
-   */
-  toText(formatter: DateFormatter = defaultDateFormatter): string {
-    const {
-      freq,
-      interval = 1,
-      count,
-      until,
-      byDay,
-      byHour,
-      byMinute,
-      byMonth,
-      byMonthDay,
-    } = this.opts;
-
-    const parts: string[] = ["every"];
-
-    const base = {
-      YEARLY: "year",
-      MONTHLY: "month",
-      WEEKLY: "week",
-      DAILY: "day",
-      HOURLY: "hour",
-      MINUTELY: "minute",
-      SECONDLY: "second",
-    }[freq];
-
-    // Special cases for WEEKLY with full weekday sets
-    const daysNormalized = byDay?.map((d) => d.toUpperCase());
-    const isWeekdays =
-      daysNormalized &&
-      daysNormalized.length === 5 &&
-      ["MO", "TU", "WE", "TH", "FR"].every((d) => daysNormalized.includes(d));
-    const isEveryday =
-      daysNormalized &&
-      daysNormalized.length === 7 &&
-      ["MO", "TU", "WE", "TH", "FR", "SA", "SU"].every((d) =>
-        daysNormalized.includes(d)
-      );
-
-    if (freq === "WEEKLY" && interval === 1 && isWeekdays) {
-      parts.push("weekday");
-    } else if (freq === "WEEKLY" && interval === 1 && isEveryday) {
-      parts.push("day");
-    } else {
-      if (interval !== 1) {
-        parts.push(interval.toString(), base + "s");
-      } else {
-        parts.push(base);
-      }
-    }
-
-    if (freq === "WEEKLY" && byDay && !isWeekdays && !isEveryday) {
-      parts.push("on", list(byDay, formatByDayToken));
-    } else if (byDay && freq !== "WEEKLY") {
-      parts.push("on", list(byDay, formatByDayToken));
-    }
-
-    if (byMonth) {
-      parts.push(
-        "in",
-        list(byMonth, (m) => MONTH_NAMES[(m as number) - 1]!)
-      );
-    }
-
-    if (byMonthDay) {
-      parts.push(
-        "on the",
-        list(byMonthDay, (d) => ordinal(d as number)),
-        "day of the month"
-      );
-    }
-
-    if (byHour) {
-      const minutes = byMinute ?? [0];
-      const times = byHour.flatMap((h) => minutes.map((m) => formatTime(h, m)));
-      parts.push("at", list(times));
-      parts.push(tzAbbreviation(this.originalDtstart));
-    }
-
-    if (until) {
-      parts.push(
-        "until",
-        formatter(until.year, MONTH_NAMES[until.month - 1]!, until.day)
-      );
-    } else if (count !== undefined) {
-      parts.push("for", count.toString(), count === 1 ? "time" : "times");
-    }
-
-    return parts.join(" ");
-  }
   /**
    * Given any date in a month, return all the ZonedDateTimes in that month
    * matching your opts.byDay and opts.byMonth (or the single "same day" if no BYDAY).
